docker-template/requirements.txt
h5py==2.7.1
numpy==1.13.3
pandas==0.21.0
Pillow==4.3.0
matplotlib==2.0.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
